Remove htmlutil
This adds the goquery dependency, which is rather large. It's annoying
in projects where you never use this function (that is, most projects)
and where you just want access to 3 or 4 functions.

The "Rob Pike method"™ is to copy those functions, but meh; I'd rather
use this package with tested functions.