Move mathutil.ComplementsInt to sliceutil
There was already sliceutil.Difference, and sliceutil makes more sense,
IMHO.

Also change the implementation to preserve the order of the list; IMHO
changing the order of any list is rather unexpected. This is slower, but
still plenty of past enough for most cases.

If you need performance, just use a map as a set from the start :-)