Add goutil (#26)
Extracted from Kommentaar. Surprisingly, there isn't a standard function
you can call for this. All this logic is "hidden" inside the
cmd/go/internal/load package (which we can't use, as it's internal).

This will also be useful in e.g. godocgen (use `go list` for that, which
works but is ugly).